backend/requirements.txt
Flask
SQLAlchemy
Flask-Script
PyJWT
cryptography
argon2-cffi